debug

+++ MessageBuilder.py
@@ -35,7 +35,7 @@ class MessageBuilder:
         Construct a message to handle communication code 3
         :return: A string describing the time and location of the object in a previous snapshot
         """
-        if int(br.location_time_passed) > 60:
+        if float(br.location_time_passed) > 60:
             message = "I seen a %s by a %s, at %s" % (br.locations_identified[0].object, br.locations_identified[0].location, br.location_time)
         else:
             message = "I seen a %s by a %s %s minutes ago" % (br.locations_identified[0].object, br.locations_identified[0].location, br.location_time_time_passed)
@@ -51,7 +51,7 @@ class MessageBuilder:
         message = "I seen a %s by a %s " % (br.locations_identified[0].object, br.locations_identified[0].location)
         for location in br.locations_identified[1:]:
             message += "and one by a %s." % (location.location)
-        if int(br.location_time_passed) > 60:
+        if float(br.location_time_passed) > 60:
             message += "at %s" % (br.location_time)
         else:
             message += "%s minutes ago" % (br.location_time_passed)
